api/server.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="Physical AI & Humanoid Robotics Textbook API",
    description="API for the AI-native textbook platform teaching Physical AI and Humanoid Robotics",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # In production, replace with specific origins
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Data models based on data-model.md
class Chapter(BaseModel):
    id: str
    title: str
    number: int
    objectives: List[str]
    lessons: List['Lesson']
    personalization_options: Optional[dict] = None
    urdu_translation_available: bool = False

class Lesson(BaseModel):
    id: str
    title: str
    chapter_id: str
    number: int
    content: str
    concepts: List[str]
    code_examples: List[dict] = []
    simulation_exercises: List[dict] = []
    mental_models: List[dict] = []
    urdu_translation_available: bool = False

class ChatQuery(BaseModel):
    query: str
    context: Optional[str] = None
    userId: Optional[str] = None

class ChatResponse(BaseModel):
    response: str
    sources: List[str]
    confidence: float

class UserProfile(BaseModel):
    id: str
    software_experience: str
    robotics_experience: str
    available_hardware: str
    preferred_language: str = "English"
    personalization_enabled: bool = True

# Sample data (in a real implementation, this would come from a database)
chapters_data = []
lessons_data = []

# API Routes
@app.get("/")
def read_root():
    return {"message": "Physical AI & Humanoid Robotics Textbook API"}

@app.get("/chapters")
def get_chapters():
    # In a real implementation, this would fetch from a database
    return {"chapters": []}

@app.get("/chapters/{chapter_id}")
def get_chapter(chapter_id: str):
    # In a real implementation, this would fetch from a database
    raise HTTPException(status_code=404, detail="Chapter not found")

@app.get("/lessons/{lesson_id}")
def get_lesson(lesson_id: str):
    # In a real implementation, this would fetch from a database
    raise HTTPException(status_code=404, detail="Lesson not found")

@app.post("/chatbot/query")
def query_chatbot(chat_query: ChatQuery):
    # In a real implementation, this would connect to the RAG system
    return ChatResponse(
        response="This is a placeholder response. The actual RAG system would generate this based on textbook content.",
        sources=[],
        confidence=0.8
    )

@app.post("/users/profile")
def create_update_profile(user_profile: UserProfile):
    # In a real implementation, this would save to a database
    return user_profile

@app.get("/users/{user_id}/progress")
def get_user_progress(user_id: str):
    # In a real implementation, this would fetch from a database
    return {"progress": {}}

@app.put("/users/{user_id}/progress")
def update_user_progress(user_id: str):
    # In a real implementation, this would update in a database
    return {"message": "Progress updated successfully"}

# Health check endpoint
@app.get("/health")
def health_check():
    return {"status": "healthy"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "server:app",
        host="0.0.0.0",
        port=int(os.getenv("DOCUSAURUS_PORT", 8000)),
        reload=True
    )