requirements.txt
# AI/ML Components for Physical AI & Humanoid Robotics Textbook
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.4
openai==1.54.4
qdrant-client==1.12.1
python-dotenv==1.0.1
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
numpy==2.1.3
pandas==2.2.3
tiktoken==0.8.0
sentence-transformers==3.1.1
torch==2.5.1
transformers==4.46.3
datasets==3.1.0
accelerate==1.1.1
bitsandbytes==0.44.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-slugify==8.0.1
aiofiles==23.2.1
jinja2==3.1.2

pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-slugify==8.0.1
aiofiles==23.2.1
jinja2==3.1.2
